requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
bitarray==2.8.2
black==23.9.1
blinker==1.6.3
cached-property==1.5.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cytoolz==0.12.2
eth-abi==4.2.1
eth-account==0.9.0
eth-bloom==2.0.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.0
eth-utils==2.2.1
exceptiongroup==1.1.3
flask==3.0.0
frozenlist==1.4.0
gunicorn==21.2.0
hexbytes==0.3.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
lru-dict==1.2.0
MarkupSafe==2.1.3
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
parsimonious==0.9.0
pathspec==0.11.2
pkgutil-resolve-name==1.3.10
platformdirs==3.11.0
pluggy==1.3.0
protobuf==4.24.4
py-ecc==6.0.0
pycryptodome==3.19.0
pytest==7.4.2
python-dotenv==1.0.0
pyunormalize==15.0.0
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rlp==3.0.0
rpds-py==0.10.4
semantic-version==2.10.0
sortedcontainers==2.4.0
tomli==2.0.1
toolz==0.12.0
trie==2.1.1
typing-extensions==4.8.0
urllib3==2.0.6
web3==6.10.0
websockets==11.0.3
werkzeug==3.0.0
yarl==1.9.2
zipp==3.17.0
